Move run to a seperate file
